refactored Configurator
