+++ src/component/mxgraph/MxGraphConfigurator.ts
@@ -1,6 +1,7 @@
 import { mxgraph } from 'ts-mxgraph';
-import { ShapeBpmnElementKind } from '../../model/bpmn/shape/ShapeBpmnElementKind';
 import { MxGraphFactoryService } from '../../service/MxGraphFactoryService';
+import StyleConfigurator from './StyleConfigurator';
+import ShapeConfigurator from './ShapeConfigurator';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 /**
@@ -13,177 +14,25 @@ export default class MxGraphConfigurator {
   private mxClient: any = MxGraphFactoryService.getMxGraphProperty('mxClient');
   private mxGraph: any = MxGraphFactoryService.getMxGraphProperty('mxGraph');
   private mxGraphModel: any = MxGraphFactoryService.getMxGraphProperty('mxGraphModel');
-  private mxUtils: any = MxGraphFactoryService.getMxGraphProperty('mxUtils');
-  private mxConstants: any = MxGraphFactoryService.getMxGraphProperty('mxConstants');
-  private mxPerimeter: any = MxGraphFactoryService.getMxGraphProperty('mxPerimeter');
-  private mxShape: any = MxGraphFactoryService.getMxGraphProperty('mxShape');
 
   private readonly graph: mxgraph.mxGraph;
 
   constructor(container: Element) {
     this.initMxGraphPrototype();
     this.graph = new this.mxGraph(container, new this.mxGraphModel());
-    this.configureStyles();
+    const styleConfigurator = new StyleConfigurator(this.graph);
+    styleConfigurator.configureStyles();
+    const shapeConfigurator = new ShapeConfigurator();
+    shapeConfigurator.initMxShapePrototype(this.mxClient.IS_FF);
   }
 
   public getGraph(): mxgraph.mxGraph {
     return this.graph;
   }
 
-  private configureStyles(): void {
-    this.configureDefaultVertexStyle();
-    this.configurePoolStyle();
-    this.configureLaneStyle();
-    this.configureStartEventStyle();
-    this.configureEndEventStyle();
-    this.configureUserTaskStyle();
-    this.configureTaskStyle();
-    this.configureParallelGatewayStyle();
-    this.configureExclusiveGatewayStyle();
-  }
-
   private initMxGraphPrototype(): void {
     this.mxGraph.prototype.edgeLabelsMovable = false;
     this.mxGraph.prototype.cellsLocked = true;
     this.mxGraph.prototype.cellsSelectable = false;
-    const isFF = this.mxClient.IS_FF;
-
-    // this change is needed for adding the custom attributes that permits identification of the BPMN elements
-    this.mxShape.prototype.createSvgCanvas = function() {
-      const mxSvgCanvas2D: any = MxGraphFactoryService.getMxGraphProperty('mxSvgCanvas2D');
-      const canvas = new mxSvgCanvas2D(this.node, false);
-      canvas.strokeTolerance = this.pointerEvents ? this.svgStrokeTolerance : 0;
-      canvas.pointerEventsValue = this.svgPointerEvents;
-      canvas.blockImagePointerEvents = isFF;
-      const off = this.getSvgScreenOffset();
-
-      if (off != 0) {
-        this.node.setAttribute('transform', 'translate(' + off + ',' + off + ')');
-      } else {
-        this.node.removeAttribute('transform');
-      }
-
-      // add attributes to be able to identify elements in DOM
-      if (this.state && this.state.cell) {
-        this.node.setAttribute('class', 'class-state-cell-style-' + this.state.cell.style);
-        this.node.setAttribute('data-cell-id', this.state.cell.id);
-      }
-      //
-      canvas.minStrokeWidth = this.minSvgStrokeWidth;
-
-      if (!this.antiAlias) {
-        // Rounds all numbers in the SVG output to integers
-        canvas.format = function(value: string) {
-          return Math.round(parseFloat(value));
-        };
-      }
-
-      return canvas;
-    };
-  }
-
-  private getStylesheet(): any {
-    return this.graph.getStylesheet();
-  }
-
-  private getDefaultVertexStyle(): any {
-    return this.getStylesheet().getDefaultVertexStyle();
-  }
-
-  private cloneDefaultVertexStyle(): any {
-    const defaultStyle = this.getDefaultVertexStyle();
-    return this.mxUtils.clone(defaultStyle);
-  }
-
-  private putCellStyle(name: ShapeBpmnElementKind, style: any): void {
-    this.getStylesheet().putCellStyle(name, style);
-  }
-
-  private configureDefaultVertexStyle(): void {
-    const style = this.getDefaultVertexStyle();
-    style[this.mxConstants.STYLE_HORIZONTAL] = true;
-    style[this.mxConstants.STYLE_FONTSIZE] = 15;
-    style[this.mxConstants.STYLE_FILLCOLOR] = 'white';
-    style[this.mxConstants.STYLE_FONTCOLOR] = 'black';
-    style[this.mxConstants.STYLE_STROKECOLOR] = 'black';
-    style[this.mxConstants.STYLE_LABEL_BACKGROUNDCOLOR] = 'none';
-  }
-
-  private configureLaneStyle(): void {
-    const style = this.cloneDefaultVertexStyle();
-    style[this.mxConstants.STYLE_SHAPE] = this.mxConstants.SHAPE_SWIMLANE;
-    style[this.mxConstants.STYLE_VERTICAL_ALIGN] = 'middle';
-    style[this.mxConstants.STYLE_HORIZONTAL] = false;
-    style[this.mxConstants.STYLE_SWIMLANE_LINE] = 0; // hide the line between the title region and the content area
-
-    this.graph.getStylesheet().putCellStyle(ShapeBpmnElementKind.LANE, style);
-  }
-
-  private configurePoolStyle(): void {
-    const style = this.cloneDefaultVertexStyle();
-    style[this.mxConstants.STYLE_SHAPE] = this.mxConstants.SHAPE_SWIMLANE;
-    style[this.mxConstants.STYLE_VERTICAL_ALIGN] = 'middle';
-    style[this.mxConstants.STYLE_HORIZONTAL] = false;
-    style[this.mxConstants.STYLE_FILLCOLOR] = '#d3d2d1';
-
-    // TODO manage pool text area rendering
-    // most of BPMN pool are ok when setting it to 30
-    style[this.mxConstants.STYLE_STARTSIZE] = 30;
-
-    this.graph.getStylesheet().putCellStyle(ShapeBpmnElementKind.POOL, style);
-  }
-
-  private configureStartEventStyle(): void {
-    const style = this.cloneDefaultVertexStyle();
-    style[this.mxConstants.STYLE_SHAPE] = this.mxConstants.SHAPE_ELLIPSE;
-    style[this.mxConstants.STYLE_PERIMETER] = this.mxPerimeter.EllipsePerimeter;
-    style[this.mxConstants.STYLE_VERTICAL_ALIGN] = 'top';
-    style[this.mxConstants.STYLE_STROKECOLOR] = '#62A928';
-    style[this.mxConstants.STYLE_STROKEWIDTH] = 1.7;
-    style[this.mxConstants.STYLE_GRADIENTCOLOR] = '#E9ECB1';
-    this.putCellStyle(ShapeBpmnElementKind.EVENT_START, style);
-  }
-
-  private configureEndEventStyle(): void {
-    const style = this.mxUtils.clone(this.getStylesheet().getCellStyle(ShapeBpmnElementKind.EVENT_START), this.getDefaultVertexStyle());
-    style[this.mxConstants.STYLE_STROKECOLOR] = '#8A151A';
-    style[this.mxConstants.STYLE_GRADIENTCOLOR] = 'Crimson';
-    this.putCellStyle(ShapeBpmnElementKind.EVENT_END, style);
-  }
-
-  private configureUserTaskStyle(): void {
-    const style = this.cloneDefaultVertexStyle();
-    style[this.mxConstants.STYLE_SHAPE] = this.mxConstants.SHAPE_RECTANGLE;
-    style[this.mxConstants.STYLE_VERTICAL_ALIGN] = 'middle';
-    style[this.mxConstants.STYLE_STROKECOLOR] = '#2C6DA3';
-    style[this.mxConstants.STYLE_STROKEWIDTH] = 2;
-    style[this.mxConstants.STYLE_ROUNDED] = true;
-    this.putCellStyle(ShapeBpmnElementKind.TASK_USER, style);
-  }
-
-  private configureTaskStyle(): void {
-    const style = this.mxUtils.clone(this.getStylesheet().getCellStyle(ShapeBpmnElementKind.TASK_USER), this.getDefaultVertexStyle());
-    style[this.mxConstants.STYLE_STROKECOLOR] = '#663399';
-    this.putCellStyle(ShapeBpmnElementKind.TASK, style);
-  }
-
-  private configureParallelGatewayStyle(): void {
-    const style = this.cloneDefaultVertexStyle();
-    style[this.mxConstants.STYLE_SHAPE] = this.mxConstants.SHAPE_RHOMBUS;
-    style[this.mxConstants.STYLE_PERIMETER] = this.mxPerimeter.RhombusPerimeter;
-    style[this.mxConstants.STYLE_VERTICAL_ALIGN] = 'top';
-    style[this.mxConstants.STYLE_STROKECOLOR] = '#96A826';
-    style[this.mxConstants.STYLE_STROKEWIDTH] = 1.7;
-
-    style[this.mxConstants.STYLE_SPACING_TOP] = 55;
-    style[this.mxConstants.STYLE_SPACING_RIGHT] = 110;
-    style[this.mxConstants.STYLE_GRADIENTCOLOR] = '#E9ECB1';
-    this.putCellStyle(ShapeBpmnElementKind.GATEWAY_PARALLEL, style);
-  }
-
-  private configureExclusiveGatewayStyle(): void {
-    const style = this.mxUtils.clone(this.getStylesheet().getCellStyle(ShapeBpmnElementKind.GATEWAY_PARALLEL), this.getDefaultVertexStyle());
-    style[this.mxConstants.STYLE_GRADIENTCOLOR] = '#DDA0DD';
-    this.putCellStyle(ShapeBpmnElementKind.GATEWAY_EXCLUSIVE, style);
   }
 }
