+++ src/component/mxgraph/ShapeConfigurator.ts
@@ -1,0 +1,41 @@
+import { MxGraphFactoryService } from '../../service/MxGraphFactoryService';
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export default class ShapeConfigurator {
+  private mxShape: any = MxGraphFactoryService.getMxGraphProperty('mxShape');
+
+  public initMxShapePrototype(isFF: boolean): void {
+    // this change is needed for adding the custom attributes that permits identification of the BPMN elements
+    this.mxShape.prototype.createSvgCanvas = function() {
+      const mxSvgCanvas2D: any = MxGraphFactoryService.getMxGraphProperty('mxSvgCanvas2D');
+      const canvas = new mxSvgCanvas2D(this.node, false);
+      canvas.strokeTolerance = this.pointerEvents ? this.svgStrokeTolerance : 0;
+      canvas.pointerEventsValue = this.svgPointerEvents;
+      canvas.blockImagePointerEvents = isFF;
+      const off = this.getSvgScreenOffset();
+
+      if (off != 0) {
+        this.node.setAttribute('transform', 'translate(' + off + ',' + off + ')');
+      } else {
+        this.node.removeAttribute('transform');
+      }
+
+      // add attributes to be able to identify elements in DOM
+      if (this.state && this.state.cell) {
+        this.node.setAttribute('class', 'class-state-cell-style-' + this.state.cell.style);
+        this.node.setAttribute('data-cell-id', this.state.cell.id);
+      }
+      //
+      canvas.minStrokeWidth = this.minSvgStrokeWidth;
+
+      if (!this.antiAlias) {
+        // Rounds all numbers in the SVG output to integers
+        canvas.format = function(value: string) {
+          return Math.round(parseFloat(value));
+        };
+      }
+
+      return canvas;
+    };
+  }
+}
